requirements.txt
dnspython==2.1.0
python-dotenv